Initialized repository: added all project files to repository
This commit initializes the full repository with all the files for my
Machine Learning project to classify gender according to OkCupid profiles.
Note that the profiles are contained in profiles.csv.zip and need to be
unzipped before continuing.

Note, that all intermediate data generated during the analysis procedures is
ignored by default (any filename ending in *.p is ignored).